feat(hal-x86_64): store CPU-local data using `gs` (#369)
This PR adds an implementation of CPU-local data storage for x86, using
the `gs` segment to store locals. An API that's kiiiiiind of like the
`LocalKey` API from `std::thread_local!` is provided for accessing
core-local data.

In addition, I've changed `mycelium_kernel::rt` to store a reference to
the current CPU core's scheduler in its local data, so that when
`rt::spawn()` is called on a CPU core that has a running scheduler, the
task is spawned on that scheduler's local run queue, instead of on the
injector.