feat(hal-x86_64): make `alloc` a feature flag (#372)
Currently, some code in `hal-x86_64` depends on `liballoc`, while other
code does not. This makes it difficult to use parts of the HAL crate in
early boot environments without an allocator.

This commit adds an "alloc" feature flag to `hal-x86_64`, which enables
code that depends on `liballoc`.

This commit was factored out of #371.