feat(hal-x86_64): add `SystemSegment::boxed_tss` (#372)
Currently, TSS segments can only be constructed from statics. This makes
it difficult to construct a TSS dynamically when bringing up an
application processor.

This commit adds a new `SystemSegment::boxed_tss` constructor, which
allows TSS segments to be dynamically allocated, as well as stored in a
static.

This was factored out of PR #371.