feat(hal-x86_64): add `Msr::ia32_efer` and `Efer` bitflags (#372)
This commit adds support for the IA32_EFER model-specific register in
the `hal-x86_64` crate's `cpu::msr` module. This includes a new `Efer`
bitflags type representing the flags stored in the EFER register, and
a `Msr::i32_efer` constructor for accessing the EFER.

This is necessary for application processor bringup, and was factored
out from PR #371.