feat(maitake): add `Wait::{waits_on, same_queue}`
This branch adds new methods to the `Wait` future returned by
`WaitQueue`s for testing whether a `Wait` future is waiting on a
particular queue instance.

`Wait::waits_on` takes a `&WaitQueue`, and returns `true` if the `Wait`
future came from that queue, while `Wait::same_queue` takes another
`&Wait` and returns `true` if the two `Wait` futures are waiting on the
same `WaitQueue` or not.

I just thought these could potentially be nice to have.