docs(maitake): document use of `portable-atomic` (#384)
This commit adds a section to the README discussing the use of the
`portable-atomic` crate. In particular, this section discusses how the
`--cfg portable_atomic_unsafe_assume_single_core` `RUSTFLAGS` cfg must
be set when compiling for targets which lack atomics entirely but are
not always guaranteed to be single-core only.