fix(hal-x86_64): ensure APIC MMIO pages are mapped (#406)
When booting using UEFI, it is possible that the I/O APIC and local APIC
memory-mapped IO pages may not be mapped or set as writable. This
results in a page fault when attempting to configure the APIC.

This commit changes the code for constructing `IoApic` and `LocalApic`
handles to take a page mapper and page-frame allocator, so that we can
ensure that the APIC MMIO pages are mapped before attempting to access
them.