feat(maitake): add `subscribe` to `WaitQueue` futures (#410)
This method allows eagerly subscribing to wakeups from a `WaitQueue`
before a future is polled for the first time. It is similar to the
`Notified::enable` method on `tokio::sync::Notify`'s `Notified` future,
but returns a `Poll<Result<(), Closed>>` rather than a `bool` in order
to indicate whether the wakeup was due to the wait queue being closed.