fix(hal-x86_64): fix race in interrupt controller init
The interrupt controller is stored in an `InitOnce` cell, so that it may
be accessed in interrupt handlers to send an EOI on completion of an
interrupt. Currently, the `enable_hardware_interrupts` function will
execute an `sti` instruction (enabling interrupts) *before* it
initializes the interrupt controller's `InitOnce` cell. This means that
if an interrupt occurs immediately after the `sti`, it can cause a race
where the static is accessed before it's initialized.

Fixes #383