feat(hal-core): pass page frame allocator immutably (#417)
The `page::Alloc` trait's methods take an `&self` receiver, so page
frames can be allocated with a shared reference (allocator impls
typically perform internal synchronization). However, the `page::Map`
trait's methods that require a page frame allocator take an `&mut`
reference to the frame allocator. This is an artifact of an earlier
version of the `page::Alloc` trait whose methods had a mutable `&mut
self` receiver, but is no longer necessary.

This commit changes the `page::Map` trait to take a shared reference to
the page-frame allocator when one is required.