feat(bitfield): select Unicode format with `Display` alt-mode (#420)
This commit changes the generated bitfield types to default to using the
ASCII `fmt::Display` multi-line format, with the Unicode format
selectable using the alt-mode formatting specifier (`{:#}`). This is to
make it easier for downstream callers to select which format is used.

The `display_ascii()` method can be used to select ASCII only regardless
of a downstream caller's format specifier, and a similar
`display_unicode()` method is added to guarantee Unicode formatting.

I don't think this should be considered a breaking change, since
existing code will still compile, and I don't consider the output of
human-readable formats not intended for machine consumption to be part
of the semver API surface. But, it will change the default behavior...