feat(maitake): add local (`!Send`) scheduler types
Currently, `maitake`'s scheduler types require tasks to be `Send`. This
is because the scheduler itself is `Send` and `Sync`, and it may be
ticked from more than one thread/CPU core. Ticking a scheduler which
contains `!Send` tasks spawned from one thread on another thread is
essentially equivalent to transferring ownership of those `!Send` tasks
across threads, so a `Send + Sync` scheduler cannot execute `!Send`
futures.

This branch introduces new `LocalScheduler` and `LocalStaticScheduler`
types for executing `!Send` futures. Other than the fact that they can
spawn futures which are not `Send`, and are themselves not `Send` or
`Sync`, they are functionally the same as the existing `Scheduler` and
`StaticScheduler` types. Additionally, in order to support other threads
to spawn tasks which *are* `Send` on a local scheduler instance, new
`LocalSpawner` and `LocalStaticSpawner` types can be created and shared
across thread boundaries. These spawner types can *only* spawn `Send`
tasks on the scheduler, and are not capable of ticking the scheduler.
Therefore, they cannot accidentally access a `!Send` future.