test(maitake): properly capture tracing messages (#445)
The "init tracing" test helper returned a guard that the caller
is supposed to hold, but none (or not all?) of the tests that called it
actually held it.

This makes a single return type (because `rustc` was Big Mad when I used
`Box<dyn Drop>`) based on the features, and marks it as must use.

This change allows for capturing (or viewing) of tracing output during
timer wheel tests.

Closes #444