feat(maitake): introduce a separate `maitake-sync` crate (#462)
The `maitake` crate provides async synchronization primitives in a
`maitake::sync` module. These APIs don't require the rest of the
runtime, and would be nice to be able to use in other crates. `maitake`
is not yet published to crates.io, as it contains Git-only dependencies.
However, the `sync` module can be implemented entirely using published
code, and therefore, a crate containing only `maitake::sync` could be
published.

This branch introduces a new `maitake-sync` crate, containing the
synchronization primitives from `maitake::sync`. The `maitake::sync`
module now re-exports `maitake-sync`, so there should be no impact for
existing uses of the crate. The new `maitake-sync` crate can easily be
published to crates.io, and depended on from released code.