feat(maitake): remove `now` from `Timer::advance` (#475)
Now that we have a trait for hardware clocks, we expect the hardware
timestamp to always come from the clock, rather than being passed in
when the timer wheel is turned. This commit changes the timer APIs so
that, rather than advancing the timer wheel to a particular `now` time,
we instead call `Timer::turn`, which advances the wheel to "whatever the
`Clock` thinks the current time is". This also means we can get rid of
`pending_ticks`, which I always kinda hated.