feat(maitake): turn timer when inserting `Sleep`s (#475)
This commit changes the `Sleep` future in `maitake::time` to turn the
timer to the `Clock`'s current `now()` time when the timer wheel lock is
acquired to register a new `Sleep`. This way, the timer is (at least
occasionally) turned while polling `Sleep`s, in addition to when
scheduled by an interrupt or by a scheduler tick completing. This way,
we fire completing sleeps much more frequently, which should help
improve the timer's accuracy substantially. Because we only do this when
the timer wheel lock already has to be acquired to register the `Sleep`
in the first place, this shouldn't increase contention for the wheel
lock all that much.

Fixes #474