chore(maitake): actually run loom tests on CI (#483)
It turns out CI only ever runs Miri tests for `cordyceps`, not
`maitake`/`maitake-sync`. MY BAD LOL.

This PR fixes that. In doing so, I've also had to go through and change
a bunch of tests to spawn fewer tasks when running under Miri, so that
it doesn't take a really long time to run. Most of the issues we hope
Miri catches don't actually depend on concurrency --- we have `loom`
tests for that.