chore(maitake): use std `pin!`, `poll_fn`, `ready`
Currently, we use the versions of `pin_mut!`, `future::poll_fn`, and
`future::ready` from the `futures` crate. Now, `core::pin::pin!`,
`core::future::poll_fn`, and `core::future::ready` are stable, so we can
are stable, so we can just use the stdlib versions of these APIs
instead.

This allows us to reduce our dev dependencies on the `futures` crates
down to just depending on `futures_util` in `maitake-sync` for
`select_biased!`. Reducing the number of dev-deps will hopefully make
compiling the tests a little faster. And, the stdlib version of `pin!`
is *way* nicer than `futures::pin_mut!` as it can be on the right-hand
side of an assignment, rather than requiring an assignment followed by a
`pin!`.