fix(hal-x86_64): don't log every MSR read (#499)
Currently, we have some code for reading MSRs that emits a tracing event
for every `rdmsr` instruction you execute. This was a stupid thing to
do, and I only did it because I was being excessively cutesy. Now that
we read the `FS_GS_BASE` MSR in ISRs to access local data, this can
deadlock the kernel :upside_down:

So, this commit removes it.