fix(hal-x86_64): remove tracing from PIT sleep (#499)
The current code for sleeping using the PIT one-shot timer has a bunch
of tracing in it. This adds overhead which makes the sleep duration way
less accurate, making PIT calibration for the local APIC timer way
sloppier. This commit removes the tracing. PIT calibration accuracy
still sucks, but it sucks less now.