fix(hal-x86_64): don't stop pagetable walk on huge tables (#502)
Currently the code for finding or creating the next-level page table
for a page will bail out if the entry pointing to the page table is a
huge page. This is because we don't currently support *mapping* huge
pages. But, if a present page table entry set up by the bootloader
happens to be a huge page, that's fine and we can still twiddle the
mappings on that table.

This fixes the kernel crashing with the latest `bootloader` crate.

Fixes #501