feat(hal-x86_64): avg multiple PIT LAPIC calibrations (#505)
This commit changes the code for calibrating the local APIC timer
frequency using the PIT do multiple PIT sleeps and average the delta
across them. This should maybe make the PIT calibration mechanism more
resilient to random noise? Eventually, we can also check if an attempt
is way outside the expected range and throw it out.