fix(inoculate): bigger test run timeout (#511)
Currently, the inoculate test command sets a timeout of 60 seconds, and
abruptly kills QEMU if it takes longer than that. This is no longer
nearly long enough to run our tests. This commit increases the timeout
to 1200 seconds (20 minutes). This is a long time, but our tests take a
while to run, and recall that this timeout applies to booting the kernel
*and running all the tests*, rather than being enforced on a per-test
basis.

We should probably make the timeout much smarter than just "kill the
QEMU process after 20 minutes", and instead do it only if we don't see
any interesting test output, or something. But, this is better than just
always failing test runs because we're stupid about timeouts, I guess.