fix(trace): Fixed infinite line-wrapping
There was a fun bug where if the offset in the line-wrapping was ever 0, it would never force a hard-wrap at the limit and also set the line back to itself in the slice.

As such it would just hopelessly spin forever trying its best to wrap a line.

This bug would also happen if the character that landed *right* on the end of the wrapping boundary was whitespace, causing the offset to be 0.