feat(cordyceps): Initial support for non-CAS targets (#526)
This PR is a first attempt to allow building of `cordyceps` on non-CAS
enabled targets, such as `thumbv6m-none-eabi`/Cortex-M0+/rp2040.

Doing this as a first feature, prior to implementing a non-atomic and
`critical_section` based version of TransferStack, in service of
https://github.com/embassy-rs/embassy/pull/4035.

Rather than making a default-on "CAS" feature, we instead gate on
`cfg(target_has_atomic = "ptr")`, which was [stabilized in
1.60](https://github.com/rust-lang/rust/pull/93824), which currently has
the meaning of "this target has CAS". This automatically gates-out
`TransferStack` and `MpscQueue`, while the "mutable" collections can
still be used.