feat(cordyceps): Add "raw" iter for `List` (#533)
This is an implementation of a "raw" iterator, that allows for
type-punning of nodes of a `List`. Unlike `MpscQueue` and
`TransferStack`, which give back `T::Handle` elements, the `iter()` and
`iter_mut()` methods of `List` instead give `&T` or `Pin<&mut T>`.
However, by creating these references, we narrow the provenance of the
reference to only the header of a type-punned object, leading to miri
stacked borrow violations.

This "RawIter" just gives "raw" `NonNull<T>` pointers to each
element, meaning that no intermediary reference is created. This is
spicy, but access to the elements still requires `unsafe` to dereference
the `NonNull`s, and the iterator itself is still bound to the lifetime
of `'list`, similar to `Iter` and `IterMut`.

---------

Co-authored-by: Eliza Weisman <eliza@buoyant.io>