fix(maitake-sync): no-cas maitake-sync (#538)
There is somewhat confusingly both `loom::sync::atomic` and
`loom::atomic`, only the latter of which uses `portable-atomic`. This
leads to maitake-sync not building on no-std crates.

This PR fixes this, and adds a test to ensure we do not regress
maitake-sync in this regard.

~We may want to do further cleanup of how we do loom exports, but this
is fine (if not still confusing) for now.~

I changed this to just stick `loom::atomic` inside `loom::sync`, so now
`loom::sync::atomic` does what I think the original intent was.

---------

Co-authored-by: Eliza Weisman <eliza@elizas.website>