doc(cordyceps): improve List rustdoc (#554)
I propose a change to the `List` rustdoc example, imo it's better
semantical match for what the `into_ptr` and `from_ptr` methods do -
which is ownership transfer via `into` -> `from` pipeline, rather than
"leaking" the ptr and then reclaiming it back again.

My `LinkedHashMap` implementation uses it, I've run it through both
`libfuzzer` as well as `miri` with no complaints, so I'd assume that
it's both sound as well as it matches the current rust memory model.